Do not use eval() to convert unknown types
The qwen3_coder tool parser included a fallback to using Python's
`eval()` to attempt to convert an unknown type into a native Python
type. This allowed execution of arbitrary Python code, as long as you
could get that Python code to be passed as an argument to a tool call.

Fixes GHSA-79j6-g2m3-jgfw

Signed-off-by: Russell Bryant <rbryant@redhat.com>